rename class for better readability
