+++ app/src/main/java/com/cody/androidexercise/MyApplication.kt
@@ -1,11 +1,11 @@
 package com.cody.androidexercise
 
 import android.app.Application
-import com.cody.androidexercise.credential.CredentialRepository
+import com.cody.androidexercise.util.SharedPrefUtil
 
 class MyApplication: Application() {
     override fun onCreate() {
         super.onCreate()
-        CredentialRepository.init(this)
+        SharedPrefUtil.init(this)
     }
 }
