+++ app/src/main/java/com/cody/androidexercise/ui/facebook/FacebookLoginViewModel.kt
@@ -2,7 +2,7 @@ package com.cody.androidexercise.ui.facebook
 
 import android.content.Intent
 import androidx.lifecycle.ViewModel
-import com.cody.androidexercise.credential.CredentialRepository
+import com.cody.androidexercise.util.SharedPrefUtil
 import com.facebook.*
 import com.facebook.login.*
 
@@ -18,7 +18,7 @@ class FacebookLoginViewModel : ViewModel() {
                 override fun onSuccess(result: LoginResult?) {
                     val accessToken = result?.accessToken
                     if (accessToken != null && !accessToken.isExpired) {
-                        CredentialRepository.facebookToken = accessToken.token
+                        SharedPrefUtil.facebookToken = accessToken.token
                     }
                 }
